move InitUserService and UserInfoModelService to users package
