+++ modules/core/src/main/java/com/developmentontheedge/be5/modules/core/controllers/UserInfoController.java
@@ -2,7 +2,7 @@ package com.developmentontheedge.be5.modules.core.controllers;
 
 import com.developmentontheedge.be5.security.UserInfoProvider;
 import com.developmentontheedge.be5.server.services.users.UserService;
-import com.developmentontheedge.be5.server.services.UserInfoModelService;
+import com.developmentontheedge.be5.server.services.users.UserInfoModelService;
 import com.developmentontheedge.be5.server.servlet.support.JsonApiController;
 import com.developmentontheedge.be5.web.Request;
