+++ modules/core/src/main/java/com/developmentontheedge/be5/modules/core/operations/users/Login.java
@@ -7,7 +7,7 @@ import com.developmentontheedge.be5.operation.OperationResult;
 import com.developmentontheedge.be5.server.FrontendActions;
 import com.developmentontheedge.be5.server.services.users.UserService;
 import com.developmentontheedge.be5.server.operations.support.GOperationSupport;
-import com.developmentontheedge.be5.server.services.UserInfoModelService;
+import com.developmentontheedge.be5.server.services.users.UserInfoModelService;
 import com.developmentontheedge.beans.DPBuilder;
 
 import javax.inject.Inject;
