+++ modules/core/src/main/java/com/developmentontheedge/be5/modules/core/operations/users/Logout.java
@@ -2,7 +2,7 @@ package com.developmentontheedge.be5.modules.core.operations.users;
 
 import com.developmentontheedge.be5.server.FrontendActions;
 import com.developmentontheedge.be5.server.operations.support.GOperationSupport;
-import com.developmentontheedge.be5.server.services.UserInfoModelService;
+import com.developmentontheedge.be5.server.services.users.UserInfoModelService;
 import com.developmentontheedge.be5.server.services.users.UserService;
 import com.developmentontheedge.be5.web.Response;
