+++ server/src/main/java/com/developmentontheedge/be5/server/ServerModule.java
@@ -17,9 +17,9 @@ import com.developmentontheedge.be5.server.services.FilterHelper;
 import com.developmentontheedge.be5.server.services.FilterInfoPlugin;
 import com.developmentontheedge.be5.server.services.FormGenerator;
 import com.developmentontheedge.be5.server.services.HtmlMetaTags;
-import com.developmentontheedge.be5.server.services.InitUserService;
+import com.developmentontheedge.be5.server.services.users.InitUserService;
 import com.developmentontheedge.be5.server.services.MenuHelper;
-import com.developmentontheedge.be5.server.services.UserInfoModelService;
+import com.developmentontheedge.be5.server.services.users.UserInfoModelService;
 import com.developmentontheedge.be5.server.services.events.EventManager;
 import com.developmentontheedge.be5.server.services.events.LogBe5Event;
 import com.developmentontheedge.be5.server.services.impl.DocumentGeneratorImpl;
