+++ server/src/main/java/com/developmentontheedge/be5/server/services/impl/InitUserServiceImpl.java
@@ -1,6 +1,6 @@
 package com.developmentontheedge.be5.server.services.impl;
 
-import com.developmentontheedge.be5.server.services.InitUserService;
+import com.developmentontheedge.be5.server.services.users.InitUserService;
 
 public class InitUserServiceImpl implements InitUserService
 {
