+++ server/src/main/java/com/developmentontheedge/be5/server/services/users/InitUserService.java
@@ -1,4 +1,4 @@
-package com.developmentontheedge.be5.server.services;
+package com.developmentontheedge.be5.server.services.users;
 
 public interface InitUserService
 {
