+++ server/src/test/java/com/developmentontheedge/be5/test/mocks/InitUserServiceMock.java
@@ -1,6 +1,6 @@
 package com.developmentontheedge.be5.test.mocks;
 
-import com.developmentontheedge.be5.server.services.InitUserService;
+import com.developmentontheedge.be5.server.services.users.InitUserService;
 
 public class InitUserServiceMock implements InitUserService
 {
